refactor(config): move longOption() to Config and update callers\n\n- Add Config.longOption(SubConfig,String) with data-type guard\n- Replace OptionUtils usages in DATASTORE_SIZE and FirstTimeWizardNewToadlet\n- Remove OptionUtils class
