+++ src/main/java/network/crypta/config/Config.java
@@ -71,4 +71,14 @@ public class Config {
   public SubConfig createSubConfig(String subConfig) {
     return new SubConfig(subConfig, this);
   }
+
+  /** Helper to obtain a numeric (long) option from a {@link SubConfig} by key. */
+  @SuppressWarnings("unchecked")
+  public static Option<Long> longOption(SubConfig subConfig, String key) {
+    Option<?> opt = subConfig.getOption(key);
+    if (opt.getDataType() != Option.DataType.NUMBER) {
+      throw new ClassCastException("Option '" + key + "' is not numeric: " + opt.getDataType());
+    }
+    return (Option<Long>) opt;
+  }
 }
